+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [Unreleased]
 ### Added
 - AGENTS.md instruction to coding agents for how to use TableTest
+- TableParser supports keeping string quotes to support table formatting tools
 
 ## [0.5.8] - 2025-12-15
 ### Fixed